fix: eliminate last 2 circular dependencies (pipeline↔state-store, telegram types↔topic-persistence)
- pipeline↔state-store: extract PipelineState to pipeline-types.ts
- telegram/types↔topic-persistence: extract SessionTopic to session-topic-types.ts

Previous PRs (#4488, #4492) already fixed 3 of 5. This eliminates the remaining 2.

Verified: TSC clean, madge clean, 0 circular dependencies.
